Add concurrency module
RLS is going to be a highly concurrent application, with three main
concurrency sources:

* user input/requests from the client editor
* changes from the file system
* internal parallelisation of CPU-heavy tasks

Because developing, debugging and testing concurrent applications is
hard, it makes sense to try to establish good concurrency patterns
from the start.

One concurrency anti-pattern is "fire & forget" spawning of background
tasks. This is problematic: without the ability to check if the task
is finished, you have to resort to busy-waiting in tests.

`concurrency` module introduces a `ConcurrentJob` object, which is a
handle to background job, which can be used to check if the job is
running and to wait for it to finish.

The idea is that any function, which schedules some job off the
current thread, should return a `ConcurrentJob`. Jobs are stored
inside the `Jobs` table, which gives a nice overview of all currently
active tasks and can be used to wait for all active tasks to finish.

All `ConcurrentJob`s must be stored in a `Jobs` table, it's and error
to drop a job on the floor.